Port CoreCLR AssemblyName.FullName change to CoreRT
(https://github.com/dotnet/coreclr/pull/11505)

This change also splits out the formatter code so
it can be in the shared partition without
having to bring along RuntimeAssemblyName and
other CoreRT-specific stuff.
